// src/web/src/api/generated/models/CreateRaceGoalRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateRaceGoalRequest = {
    goalTimeSeconds: number | string | null;
    goalPaceSecondsPerKm: number | string | null;
    confidence: string | null;
    rationale: string;
};