// src/web/src/api/generated/models/SaveTargetRaceRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SaveTargetRaceRequest = {
    name: string;
    raceDate: string;
    distanceM: number | string;
    location: string | null;
    priority: string;
    status: string;
    timezoneName: string | null;
};